server/pymodule/requirements.txt
dlib==19.22.1
imutils==0.5.4
numpy==1.21.2
paho_mqtt==1.6.1
Pillow==9.2.0
requests==2.26.0
